% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_desc.R
\name{add_desc}
\alias{add_desc}
\title{Add description attributes to data frame columns}
\usage{
add_desc(.data, ...)
}
\arguments{
\item{.data}{A data frame.}

\item{...}{Named arguments where names are column names and values are description strings.}
}
\value{
The modified data frame with description (and possibly unit) attributes.
}
\description{
Adds a "description" attribute to specified columns. If the description contains
square brackets (e.g., "Length \verb{[cm]}"), the content inside the brackets is stored
as a "unit" attribute.
}
\examples{
\dontrun{
library(dplyr)
df <- iris \%>\%
  add_desc(Sepal.Length = "Sepal length", Sepal.Width = "Sepal width [cm]")

str(df)
}
}
