% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_unit.R
\name{add_unit}
\alias{add_unit}
\title{Add unit attributes to data frame columns}
\usage{
add_unit(.data, ...)
}
\arguments{
\item{.data}{A data frame.}

\item{...}{Named arguments where names are column names and values are unit strings.}
}
\value{
The modified data frame with unit attributes added to specified columns.
}
\description{
This function allows you to add "unit" attributes to specified columns in a data frame.
It is designed to work with the tidyverse and can be used with the pipe (\verb{\%>\%}) operator.
}
\examples{
library(dplyr)
df <- iris \%>\%
  add_unit(Sepal.Length = "cm", Sepal.Width = "cm")
}
