% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copySpByYear.R
\name{copySpByYear}
\alias{copySpByYear}
\title{Copy photos of specific species from specific year}
\usage{
copySpByYear(inDir, newDir, exifTag, species, year)
}
\arguments{
\item{inDir}{character. The directory where all the images are saved in.}

\item{newDir}{character. The directory where you would like to copy your images to.}

\item{exifTag}{character. The name of the species tag column. For images tagged in Windows Live Gallery use "Subject", and for images tagged in digiKam use "HierarchicalSubject" as the input argumenets. Actually, now after I have cheanged the code to include digiKam, I think "Subject" might just work for both... *palm to face}

\item{species}{character. The species tag as tagged in Windows Live Gallery.}

\item{year}{integer. The year from which you would like the images from.}
}
\value{
Copies all the images according to the specified arguments (species, year) to a new directory as chosen by the user.
}
\description{
Copy all images from a directory which has a certain metadata tag and from a certain year as specified by the user. This function is very much similar to the getSpeciesImages function from the camtrapR package except you can specify which year the images are from and it can use Windows Live Gallery tags as well as digiKam tags.
}
\examples{
copySpByYear(inDir   = "/Users/philipfaure/Desktop/testSample",
             newDir  = "/Users/philipfaure/Desktop/testOutput",
             exifTag = "Subject",
             species = "Leopard",
             year    = 2016)

}
\author{
Philip Faure, \email{philip.faure13@gmail.com}
}
\keyword{Copy}
\keyword{by}
\keyword{species}
\keyword{year}
