% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Extract Metadata}
\usage{
metadata(inDir, outDir, name)
}
\arguments{
\item{inDir}{character. The directory where your images are saved.}

\item{outDir}{character. The directory where you want your metadata saved.}

\item{name}{character. The name you want for your new metadata file.}
}
\description{
Creates a csv file with all your metadata in.
}
\examples{
metadata(inDir  = "/Users/philipfaure/Photos/",
         outDir = "/Users/philipfaure/Metadata/",
         name   = "my_metadata")
}
\author{
Philip Faure, \email{philip.faure13@gmail.com}
}
\keyword{metadata}
