% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataPPP.R
\name{metadataPPP}
\alias{metadataPPP}
\title{Extract Metadata}
\usage{
metadataPPP(inDir, outDir, name)
}
\arguments{
\item{inDir}{character. The directory where your images are saved.}

\item{outDir}{character. The directory where you want your metadata saved.}

\item{name}{character. The name you want for your new metadata file.}
}
\description{
Creates a csv file with selected metadata columns in. SourceFile, AmbientTemperature, MoonPhase, DateTimeOriginal, and Keywords are returned in the csv file.
}
\examples{
metadataPPP(inDir  = "/Users/philipfaure/Photos",
            outDir = "/Users/philipfaure/Metadata",
            name   = "my_metadata")
}
\author{
Philip Faure, \email{philip.faure13@gmail.com}
}
\keyword{metadata}
