% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_naming_convention.R
\name{view_naming_convention}
\alias{view_naming_convention}
\title{View a naming convention in a readable table}
\usage{
view_naming_convention(naming_object)
}
\arguments{
\item{naming_object}{A tibble with columns: name, mandatory, valid_values, default_description}
}
\value{
A flextable object
}
\description{
Displays the naming convention tibble using flextable formatting.
}
\examples{
ft<-view_naming_convention(EcoForest_naming)
print(ft, preview = "html")
}
