% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_desc.R
\name{add_desc}
\alias{add_desc}
\title{Add description attributes to data frame columns}
\usage{
add_desc(.data, ...)
}
\arguments{
\item{.data}{A data frame.}

\item{...}{Column descriptions, either as named arguments (e.g., Sepal.Length = "Sepal length") or as a single named vector/list (e.g., c(Sepal.Length = "Sepal length", Sepal.Width = "Sepal width \link{cm}")).}
}
\value{
The modified data frame with description (and possibly unit) attributes.
}
\description{
Adds a "description" attribute to specified columns. If the description contains
square brackets (e.g., "Length \verb{[cm]}"), the content inside the brackets is stored
as a "unit" attribute.
}
\examples{
library(dplyr)
df <- iris \%>\%
  add_desc(Sepal.Length = "Sepal length", Sepal.Width = "Sepal width `[cm]`") #cm will be stored as unit for sepal width, while no unit will be stored for sepal length

}
